Change the “Organize” behavior to leave albums unmoved if they’re already in a collection with a matching title, and move albums to the first existing collection with a matching title if possible.
Keep the empty collection undeleted and visible but disabled in the “organize albums” sheet if organizing would empty the collection. Wait until the user taps “Save” to delete it.

Change the prompt in the “organize albums” sheet to “Organize [X] albums by album artist?”

Fix a bug where you were unnecessarily allowed to organize albums with unknown album artists that were already in a collection with a matching title.

Distributed as build 297.